feat(download-images): created DownloadImagesSelect.tsx
